Added missing fn & Clearer Docs & Reorg
