+++ src/contract/ens/mod.rs
@@ -1,0 +1,8 @@
+//! Ethereum Name Service
+
+mod eth_ens;
+mod public_resolver;
+mod registry;
+mod reverse_resolver;
+
+pub use eth_ens::Ens;
