+++ src/contract/mod.rs
@@ -14,6 +14,7 @@ use crate::{
 use std::{collections::HashMap, hash::Hash, time};
 
 pub mod deploy;
+pub mod ens;
 mod error;
 pub mod tokens;
